% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_attr}
\alias{add_attr}
\title{Add attributes from one data frame to another}
\usage{
add_attr(df1, df2)
}
\arguments{
\item{df1}{Data frame to receive attributes.}

\item{df2}{Data frame to give attributes.}
}
\value{
Data frame.
}
\description{
Add "map" attributes from df2 to df1.
}
\examples{
df1 <- data.frame(cars = 1:3)
df2 <- data.frame(cars = 1:3)
attr(df2$cars, "map") <- data.frame(levels = 1:3,
labels = c("Volvo", "Saab", "Opel"))
df1 <- add_attr(df1, df2)
df1$cars
}
