% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_units.R
\name{all_units}
\alias{all_units}
\title{All active units}
\usage{
all_units(period = lubridate::year(Sys.Date()))
}
\arguments{
\item{period}{Year/years of primary operation and year/years of PROM date.}
}
\value{
List containing all units, units who register primaries, units who register reoperations and units who register postoperative PROM.
}
\description{
Get all active units for specific period.
}
\examples{
# Get all active units for 2018
all_units()
# Get all active units for 2016-2017
all_units(2016:2017)
# Get all active units for 1900-2020 (all existing years)
all_units(1900:2020)
}
