% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_to_factor.R
\name{attr_to_factor}
\alias{attr_to_factor}
\title{Convert object with attributes in "map" to factor}
\usage{
attr_to_factor(x)
}
\arguments{
\item{x}{Object with "map" attributes, where levels and labels exist}
}
\value{
Factor
}
\description{
Used to extract texts from a variable containing numeric values.
}
\examples{
df <- data.frame(cars = 1:3)
attr(df$cars, "map") <- data.frame(levels = 1:3, labels = c("Volvo", "Saab", "Opel"))
attr_to_factor(df$cars)
}
