% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_attr}
\alias{check_attr}
\title{Check if object has attribute "map"}
\usage{
check_attr(x)
}
\arguments{
\item{x}{Vector}
}
\value{
TRUE or FALSE
}
\description{
Useful in mutate_if to be used with mutate_label
}
\examples{
df <- data.frame(cars = 1:3)
attr(df$cars, "map") <-
data.frame(levels = 1:3, labels = c("Volvo", "Saab", "Opel"))
check_attr(df$cars)
}
