% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existing_prom.R
\name{existing_prom}
\alias{existing_prom}
\title{Existing PROM}
\usage{
existing_prom(period = lubridate::year(Sys.Date()))
}
\arguments{
\item{period}{Year/years of primary operation.}
}
\value{
Data frame with units, period, number of operations, number of THR, number of hemi,
number of primary OA patients, number of fracture patients,
number of elective patients and number of existing PROM.
}
\description{
Number of existing preoperative PROM and 1/6/10 years postoperative PROM.
}
\examples{
# Get number of existing PROM for 2018
existing_prom()
# Get number of existing PROM for 2016-2017
existing_prom(2016:2017)
# Get number of existing PROM for 1900-2020 (all existing years)
existing_prom(1900:2020)
}
