% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_cens}
\alias{filter_cens}
\title{Filter variable for existing data}
\usage{
filter_cens(x, y)
}
\arguments{
\item{x}{Variable to be filtered.}

\item{y}{Condition variable.}
}
\value{
Filtered variable.
}
\description{
Filter variable where there are existing data in other variable.
}
\examples{
df <- data.frame(
  cars = 1:3,
  cars_model = 4:6,
  cars_speed = 7:9,
  condition = c(NA, 1, 2)
)
dplyr::mutate_at(
  df,
  dplyr::vars(dplyr::starts_with("cars")),
  filter_cens, y = df$condition
)
}
