% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Load SLR data layer}
\usage{
get_data(form = "all", functions = FALSE)
}
\arguments{
\item{form}{Which tables to load. Either 'all', 'allhip', 'allknee', or one or several FormIDs.
'1002' HipOperation (primary + reoperations)
'2259' HipEnvironment
'2253' HipPreProm
'2254' HipPostProm
'2334' KneeOperation (primary + reoperations)
'2346' KneePreProm
'2347' KneePostProm
'2335' OsteoOperation (primary + reoperations)
'all'  All forms
'allhip' All hip forms
'allknee' All knee forms}

\item{functions}{Include functions from data layer?}
}
\description{
Load SLR data layer
}
\examples{
\dontrun{
get_data()
get_data('allhip')
get_data(form = c(1002, 2334), functions = TRUE)
}
}
