% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_units.R
\name{get_units}
\alias{get_units}
\title{Extract units from data}
\usage{
get_units(x)
}
\arguments{
\item{x}{Vector}
}
\value{
All unique units sorted in alpabetic order.
}
\description{
Get units for specific period. Only used as a help function for all_units.
}
\examples{
get_units(dataOperations$P_Unit)
}
