% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_label}
\alias{mutate_label}
\title{Extract attribute "map" and mutate object to be labelled}
\usage{
mutate_label(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Labelled
}
\description{
Labelled are used to get right labels when saving to spss using package haven
}
\examples{
df <- data.frame(cars = 1:3)
attr(df$cars, "map") <- data.frame(levels = 1:3,
labels = c("Volvo", "Saab", "Opel"))
df <- dplyr::mutate(df, dplyr::across(where(check_attr), mutate_label))
}
