% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim_op.R
\name{prim_op}
\alias{prim_op}
\title{Primary operations}
\usage{
prim_op(period = lubridate::year(Sys.Date()))
}
\arguments{
\item{period}{Year/years of primary operation.}
}
\value{
Data frame with units, period, number of operations, number of THR and number of hemi.
}
\description{
Get number of primary operations for each unit and specific period.
}
\examples{
# Get number of operations for current year
prim_op()
# Get number of operations for 2016-2017
prim_op(2016:2017)
# Get number of operations for 1900-2020 (all existing years)
prim_op(1900:2020)
}
