% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_units.R
\name{profile_units}
\alias{profile_units}
\title{Operation profile units}
\usage{
profile_units(period = lubridate::year(Sys.Date()))
}
\arguments{
\item{period}{Year/years of primary operation.}
}
\value{
List containing units with and without profile.
}
\description{
Get units with and without profile for specific period.
}
\examples{
# Get units with and without profile for 2018
profile_units()
# Get units with and without profile for 2016-2017
profile_units(2016:2017)
# Get units with and without profile for 1900-2020 (all existing years)
profile_units(1900:2020)
}
