% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_op.R
\name{re_op}
\alias{re_op}
\title{Reoperations}
\usage{
re_op(period = lubridate::year(Sys.Date()))
}
\arguments{
\item{period}{Year/years of reoperation.}
}
\value{
Data frame with units, period, number of reoperations, number of THR and number of hemi.
}
\description{
Get number of reoperations for each unit and specific period.
}
\examples{
# Get number of reoperations for current year
re_op()
# Get number of reoperations for 2016-2017
re_op(2016:2017)
# Get number of reoperations for 1900-2020 (all existing years)
re_op(1900:2020)
}
