% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_op.R
\name{revision_op}
\alias{revision_op}
\title{Revisions}
\usage{
revision_op(period = lubridate::year(Sys.Date()))
}
\arguments{
\item{period}{Year/years of revision.}
}
\value{
Data frame with units, period, number of revisions, number of THR and number of hemi.
}
\description{
Get number of revisions for each unit and specific period.
}
\examples{
# Get number of revisions for current year
revision_op()
# Get number of revisions for 2016-2017
revision_op(2016:2017)
# Get number of revisions for 1900-2020 (all existing years)
revision_op(1900:2020)
}
