% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satisfaction.R
\name{satisfaction}
\alias{satisfaction}
\title{Satisfaction PROM}
\usage{
satisfaction(period = lubridate::year(Sys.Date()))
}
\arguments{
\item{period}{Year/years of primary operation.}
}
\value{
Data frame with units, period, number of operations, number of THR, number of hemi,
existing satisfaction data, number of satisfied (answered 4 or 5) and number of unsatisfied (answered 1 or 2).
}
\description{
Number of satisfied/unsatisfied patients 1 year after primary operation for a specific period. Satisfaction associated with the hip that is operated.
}
\examples{
# Get number of satisfied/unsatisfied for current year
# (does not exist since not enough follow up time)
satisfaction()
# Get number of satisfied/unsatisfied for 2016-2017
satisfaction(2016:2017)
# Get number of satisfied/unsatisfied for 1900-2020 (all existing years)
satisfaction(1900:2020)
}
