% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.xlsx.R
\name{save.xlsx}
\alias{save.xlsx}
\title{Old xlsx saver}
\usage{
save.xlsx(file, ..., row_names = FALSE, col_names = TRUE)
}
\arguments{
\item{file}{Filename.}

\item{...}{Data frames or matrices, where each data frame will be saved into one sheet.}

\item{row_names}{If TRUE, row names are included in output.}

\item{col_names}{If TRUE, column names are included in output.}
}
\description{
Used to save xlsx files.
}
