% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xlsx.R
\name{save_xlsx}
\alias{save_xlsx}
\title{New xlsx saver}
\usage{
save_xlsx(file, ..., row_names = FALSE, col_names = TRUE, bold = TRUE)
}
\arguments{
\item{file}{Filename.}

\item{...}{Data frames or matrices, where each data frame will be saved into
one sheet.}

\item{row_names}{If TRUE, row names are included in output.}

\item{col_names}{If TRUE, column names are included in output.}

\item{bold}{If TRUE, column names are centered and bold.}
}
\description{
Used to save xlsx files.
}
