% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shpr_data.R
\name{shpr_data}
\alias{shpr_data}
\title{Load SHPR data layer}
\usage{
shpr_data(factor_format = FALSE)
}
\arguments{
\item{factor_format}{If TRUE, variables with attributes are in factor format.}
}
\description{
Load all SHPR data, with attributes or factor format
}
\examples{
shpr_data()
shpr_data(TRUE)
}
