% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_summary.R
\name{survSummary}
\alias{survSummary}
\title{Survival analysis summary}
\usage{
survSummary(surv_object, times)
}
\arguments{
\item{surv_object}{Object returned from survfit function in survival package.}

\item{times}{Time to be summarized.}
}
\value{
Data frame summarizing surv_object.
}
\description{
Summary of a survfit object with more than one model from survival package.
If there is not enough follow up time the latest results is extrapolated.
}
