% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vas_diff.R
\name{vas_diff}
\alias{vas_diff}
\title{Difference in VAS health}
\usage{
vas_diff(period = lubridate::year(Sys.Date()))
}
\arguments{
\item{period}{Year/years of primary operation.}
}
\value{
Data frame with units, period, number of operations, number of THR, number of hemi,
existing VAS data (both before and 1/6/10 years after operation),
mean difference in VAS health (between 1/6/10 years after operation and before operation)
}
\description{
Mean of the difference in VAS health after operation and before operation.
}
\examples{
# Get difference in VAS health for 2018 (does not exist since not enough follow up time)
vas_diff()
# Get difference in VAS health for 2016-2017 (only 1 year follow up exist)
vas_diff(2016:2017)
# Get difference in VAS health for 1900-2020 (all existing years)
vas_diff(1900:2020)
}
